Refactor: ProductSeason 파라미터 이름 및 타입 변경
- seasonId -> id
- Long -> Integer

Resolves: #85

+++ src/main/java/com/example/shinsekai/season/application/ProductSeasonServiceImpl.java
@@ -25,7 +25,7 @@ public class ProductSeasonServiceImpl implements ProductSeasonService {
     }
 
     @Override
-    public List<ProductSeasonListGetResponseDto> getAllProductSeasonList(Long seasonId) {
+    public List<ProductSeasonListGetResponseDto> getAllProductSeasonList(Integer seasonId) {
         return productSeasonListRepository.findAllBySeasonId(seasonId)
                 .stream().map(ProductSeasonListGetResponseDto::from).toList();
     }
